fix(inference): resolve chained generic parameter defaults
