+++ crates/biome_module_graph/src/db/queries/type_inference.rs
@@ -1802,13 +1802,17 @@ fn infer_function_return_type<'db>(
 }
 
 /// Substitutes generic return references from arguments and callback results.
-/// Type parameters left unsubstituted fall back to their declared defaults.
+/// Type parameters left unsubstituted fall back to their declared defaults;
+/// a default resolves through the substitutions accumulated before it, so a
+/// default referencing an earlier type parameter follows that parameter's
+/// argument binding or default instead of leaving the bare generic behind.
 fn infer_generic_return_type<'db>(
     db: &'db dyn ModuleDb,
     function: InferredFunction<'db>,
     mut return_ty: InferredTypeData<'db>,
     args: &[ResolvedCallArgument<'db>],
 ) -> InferredTypeData<'db> {
+    let mut substitutions: Vec<InferredTypeSubstitution<'db>> = Vec::new();
     for (parameter, arg) in function
         .parameters(db)
         .iter()
@@ -1816,13 +1820,12 @@ fn infer_generic_return_type<'db>(
     {
         let parameter_ty = parameter.ty();
         if parameter_ty.is_generic_reference(db) {
-            return_ty = return_ty.substitute_type(
-                db,
-                InferredTypeSubstitution {
-                    generic: parameter_ty,
-                    replacement: arg,
-                },
-            );
+            let substitution = InferredTypeSubstitution {
+                generic: parameter_ty,
+                replacement: arg,
+            };
+            return_ty = return_ty.substitute_type(db, substitution);
+            substitutions.push(substitution);
             continue;
         }
 
@@ -1843,6 +1846,7 @@ fn infer_generic_return_type<'db>(
             collect_callback_return_replacements(db, *parameter_return_ty, *argument_return_ty)
         {
             return_ty = return_ty.substitute_type(db, substitution);
+            substitutions.push(substitution);
         }
     }
 
@@ -1850,13 +1854,15 @@ fn infer_generic_return_type<'db>(
         if let InferredTypeData::Generic(generic) = type_parameter
             && let Some(default) = generic.default(db)
         {
-            return_ty = return_ty.substitute_type(
-                db,
-                InferredTypeSubstitution {
-                    generic: *type_parameter,
-                    replacement: default,
-                },
-            );
+            let replacement = substitutions.iter().fold(default, |ty, substitution| {
+                ty.substitute_type(db, *substitution)
+            });
+            let substitution = InferredTypeSubstitution {
+                generic: *type_parameter,
+                replacement,
+            };
+            return_ty = return_ty.substitute_type(db, substitution);
+            substitutions.push(substitution);
         }
     }
