+++ crates/biome_module_graph/tests/spec_tests_v2.rs
@@ -42,6 +42,8 @@ use salsa::plumbing::{AsId, FromId};
 mod expected_argument_inference;
 #[path = "spec_tests_v2/promises.test.rs"]
 mod promises;
+#[path = "spec_tests_v2/substitutions.test.rs"]
+mod substitutions;
 
 #[salsa::db]
 struct TestModuleDb {
@@ -1523,78 +1525,6 @@ fn test_infer_module_types_preserves_generic_interface_instantiation_parameters(
         &db,
         &fs,
     );
-}
-
-#[test]
-fn test_infer_module_types_substitutes_generic_members_through_lookup_traversal() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            interface InterfaceBase<T> {
-                value: T;
-            }
-
-            interface InterfaceBox<T> extends InterfaceBase<T> {}
-
-            class ClassBase<T> {
-                value: T;
-            }
-
-            class ClassBox<T> extends ClassBase<T> {}
-
-            interface UnionBox<T> {
-                value: T;
-            }
-
-            export function readInterfaceBox(
-                value: InterfaceBox<number>,
-            ): InterfaceBox<number> {
-                return value;
-            }
-
-            export function readClassBox(value: ClassBox<string>): ClassBox<string> {
-                return value;
-            }
-
-            export function readUnionBox(
-                value: UnionBox<string> | UnionBox<number>,
-            ): UnionBox<string> | UnionBox<number> {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-
-    let interface_box_ty =
-        inferred_function_return_ty_by_name(&db, index_module, inferred, "readInterfaceBox")
-            .expect("readInterfaceBox return type must be inferred");
-    let interface_value_ty = inferred
-        .find_member_type(&db, interface_box_ty, "value")
-        .expect("InterfaceBox<number>.value must be inferred through extends");
-    assert!(is_inferred_number(&db, interface_value_ty));
-
-    let class_box_ty =
-        inferred_function_return_ty_by_name(&db, index_module, inferred, "readClassBox")
-            .expect("readClassBox return type must be inferred");
-    let class_value_ty = inferred
-        .find_member_type(&db, class_box_ty, "value")
-        .expect("ClassBox<string>.value must be inferred through extends");
-    assert!(is_inferred_string(&db, class_value_ty));
-
-    let union_box_ty =
-        inferred_function_return_ty_by_name(&db, index_module, inferred, "readUnionBox")
-            .expect("readUnionBox return type must be inferred");
-    let union_value_ty = inferred
-        .find_member_type(&db, union_box_ty, "value")
-        .expect("UnionBox<string> | UnionBox<number> value must be inferred");
-    assert!(contains_inferred_string(&db, union_value_ty));
-    assert!(contains_inferred_number(&db, union_value_ty));
 }
 
 #[test]
@@ -4357,147 +4287,6 @@ fn test_infer_call_expression_type_selects_imported_function_declaration_overloa
 }
 
 #[test]
-fn test_infer_call_expression_type_substitutes_direct_generic_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function identity<T>(value: T): T {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let identity_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "identity")
-        .expect("identity binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, identity_ty),
-        Vec::from([InferredTypeData::Number]),
-    );
-
-    assert!(is_inferred_number(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_deduplicates_substituted_union_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function maybeString<T>(value: T): T | string {
-                return value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let maybe_string_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "maybeString")
-        .expect("maybeString binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, maybe_string_ty),
-        Vec::from([InferredTypeData::String]),
-    );
-
-    assert!(is_inferred_string(&db, call_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_deduplicates_substituted_union_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_callback_generic_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function run<T>(cb: () => T): T {
-                return cb();
-            }
-
-            export function readString(): string {
-                return "value";
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let run_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "run")
-        .expect("run binding type must be inferred");
-    let read_string_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "readString")
-        .expect("readString binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, run_ty),
-        Vec::from([inferred.resolve_type(&db, read_string_ty)]),
-    );
-
-    assert!(is_inferred_string(&db, call_ty));
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_nested_generic_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function wrap<T>(value: T): Promise<T> {
-                return Promise.resolve(value);
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let wrap_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "wrap")
-        .expect("wrap binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, wrap_ty),
-        Vec::from([InferredTypeData::Number]),
-    );
-    let InferredTypeData::InstanceOf(instance) = call_ty else {
-        panic!("wrap must return a Promise instance, got {call_ty:?}");
-    };
-
-    assert!(call_ty.is_promise_instance(&db));
-    assert!(
-        instance
-            .type_parameters(&db)
-            .iter()
-            .any(|ty| is_inferred_number(&db, *ty))
-    );
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_nested_generic_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
 fn test_infer_module_types_calls_generic_function_type_aliases() {
     let fs = MemoryFileSystem::default();
     fs.insert(
@@ -4598,41 +4387,6 @@ fn test_infer_module_types_calls_nested_generic_callable_aliases() {
 }
 
 #[test]
-fn test_infer_call_expression_type_substitutes_generic_in_function_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function wrap<T>(value: T): () => T {
-                return () => value;
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, module).expect("types must be inferred");
-    let wrap_ty = inferred_binding_ty_by_name(&db, module, inferred, "wrap")
-        .expect("wrap binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        module,
-        inferred.resolve_type(&db, wrap_ty),
-        Vec::from([InferredTypeData::Number]),
-    );
-    let InferredTypeData::Function(function) = call_ty else {
-        panic!("wrap must return a function, got {call_ty:?}");
-    };
-    let InferredReturnType::Type(return_ty) = function.return_type(&db) else {
-        panic!("nested function return type must be inferred");
-    };
-
-    assert!(is_inferred_number(&db, *return_ty));
-}
-
-#[test]
 fn test_infer_call_expression_type_preserves_shadowed_nested_function_generic() {
     let fs = MemoryFileSystem::default();
     fs.insert(
@@ -4692,101 +4446,6 @@ fn test_infer_module_types_preserves_shadowed_class_method_generic() {
         inferred.resolve_type(&db, result_ty),
         "value"
     ));
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_multiple_generics_inside_union_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            export function pair<T, U>(left: T, right: U): Promise<T | U> {
-                return Promise.resolve(left as T | U);
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let pair_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "pair")
-        .expect("pair binding type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, pair_ty),
-        Vec::from([InferredTypeData::String, InferredTypeData::Number]),
-    );
-
-    assert!(is_inferred_promise_with_type_parameter(
-        &db,
-        call_ty,
-        |ty| { contains_inferred_string(&db, ty) && contains_inferred_number(&db, ty) }
-    ));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_multiple_generics_inside_union_return_type",
-        &db,
-        &fs,
-    );
-}
-
-#[test]
-fn test_infer_call_expression_type_substitutes_generic_inside_intersection_return_type() {
-    let fs = MemoryFileSystem::default();
-    fs.insert(
-        "/src/index.ts".into(),
-        r#"
-            type WithName = {
-                name: string;
-            };
-
-            export function withName<T>(value: T): T & WithName {
-                return value as T & WithName;
-            }
-
-            export function readValue(): { value: number } {
-                return { value: 1 };
-            }
-        "#,
-    );
-
-    let db = build_js_test_module_db(&fs, &["/src/index.ts"], true);
-    let index_module = db
-        .module_for_path(Utf8Path::new("/src/index.ts"))
-        .expect("module must exist");
-    let inferred = infer_module_types(&db, index_module).expect("types must be inferred");
-    let with_name_ty = inferred_binding_ty_by_name(&db, index_module, inferred, "withName")
-        .expect("withName binding type must be inferred");
-    let read_value_ty =
-        inferred_function_return_ty_by_name(&db, index_module, inferred, "readValue")
-            .expect("readValue return type must be inferred");
-    let call_ty = infer_call_expression_type(
-        &db,
-        index_module,
-        inferred.resolve_type(&db, with_name_ty),
-        Vec::from([read_value_ty]),
-    );
-    let InferredTypeData::Object(object) = call_ty else {
-        panic!("withName call must return a normalized object, got {call_ty:?}");
-    };
-    assert_eq!(object.members(&db).len(), 2);
-
-    let name_ty = inferred
-        .find_member_type(&db, call_ty, "name")
-        .expect("normalized call return must expose WithName.name");
-    assert!(is_inferred_string(&db, name_ty));
-
-    let value_ty = inferred
-        .find_member_type(&db, call_ty, "value")
-        .expect("normalized call return must expose readValue.value");
-    assert!(is_inferred_number(&db, value_ty));
-    assert_inferred_type_snapshot(
-        "test_infer_call_expression_type_substitutes_generic_inside_intersection_return_type",
-        &db,
-        &fs,
-    );
 }
 
 #[test]
